Move babel plugin files to build-system/babel-plugins/
