+++ build-system/babel-plugins/log-module-metadata.js
