+++ build-system/babel-plugins/static-template-metadata.js
