+++ build-system/eslint-rules/no-mixed-interpolation.js
@@ -19,7 +19,7 @@ const {
   definitionFile,
   singletonFunctions,
   transformableMethods,
-} = require('../log-module-metadata.js');
+} = require('../babel-plugins/log-module-metadata.js');
 
 /**
  * @param {!Node} node
