+++ build-system/pr-check/build-targets.js
@@ -46,8 +46,8 @@ const targetMatchers = [
     func: file => {
       return (
         file == 'build-system/internal-version.js' ||
-        file == 'build-system/log-module-metadata.js' ||
-        file == 'build-system/static-template-metadata.js' ||
+        file == 'build-system/babel-plugins/log-module-metadata.js' ||
+        file == 'build-system/babel-plugins/static-template-metadata.js' ||
         file == 'build-system/compile/log-messages.js' ||
         file == 'build-system/tasks/babel-plugin-tests.js' ||
         file.startsWith('build-system/babel-plugins/')
